Refactor rpc server to separate package
